Refactor SetOperation and BtnEqual_Click methods
Refactor the SetOperation method to check if `op` is not equal to "=" and is not null before setting the background color of `lblOp` to `SystemColors.GradientActiveCaption`.

In the BtnEqual_Click method, remove the division by zero check and associated error message. Simplify the method to only set the selection start of `txtResult`, reset the operation (`op`), and change the background color of `lblOp` to `Color.AliceBlue`.